Добавлена корзина для неавторизованных пользователей.

+++ store/views.py
@@ -10,15 +10,18 @@ def store(request):
     if request.user.is_authenticated:
         customer = request.user.customer
         order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        cart_items = order.get_total_order_quantity
+        cart_total_quantity = order.get_total_order_quantity
     else:
+        cart = json.loads(request.COOKIES['cart'])
         order = {'get_total_order_price':0, 'get_total_order_quantity':0}
-        cart_items = order['get_total_order_quantity']
+        for i in cart:
+            order['get_total_order_quantity'] += cart[i]['quantity']
+        cart_total_quantity = order['get_total_order_quantity']
     
     products = Product.objects.all()
     context = {
         'products': products,
-        'cart_items': cart_items,
+        'cart_total_quantity': cart_total_quantity,
     }
     return render(request, 'store/store.html', context)
 
@@ -28,16 +31,38 @@ def cart(request):
         customer = request.user.customer
         order, created = Order.objects.get_or_create(customer=customer, complete=False)
         order_items = order.orderitem_set.all()
-        cart_items = order.get_total_order_quantity
+        cart_total_quantity = order.get_total_order_quantity
     else:
+        try:
+            cart = json.loads(request.COOKIES['cart'])
+        except:
+            cart = {}
+        
         order_items = []
         order = {'get_total_order_price':0, 'get_total_order_quantity':0}
-        cart_items = order['get_total_order_quantity']
+        for i in cart:
+            product = Product.objects.get(id=i)
+            product_total_price = product.price * cart[i]['quantity']
+            order['get_total_order_price'] += product_total_price
+            order['get_total_order_quantity'] += cart[i]['quantity']
+            item = {
+                'product': {
+                    'id': product.id,
+                    'name': product.name,
+                    'price': product.price,
+                    'image': product.image,
+                },
+                'quantity': cart[i]['quantity'],
+                'get_total_items_price': product_total_price,
+            }
+            order_items.append(item)
+        
+        cart_total_quantity = order['get_total_order_quantity']
         
     context = {
         'order_items': order_items,
         'order': order,
-        'cart_items': cart_items,
+        'cart_total_quantity': cart_total_quantity,
     }
     return render(request, 'store/cart.html', context)
 
@@ -47,16 +72,38 @@ def checkout(request):
         customer = request.user.customer
         order, created = Order.objects.get_or_create(customer=customer, complete=False)
         order_items = order.orderitem_set.all()
-        cart_items = order.get_total_order_quantity
+        cart_total_quantity = order.get_total_order_quantity
     else:
+        try:
+            cart = json.loads(request.COOKIES['cart'])
+        except:
+            cart = {}
+        
         order_items = []
         order = {'get_total_order_price':0, 'get_total_order_quantity':0}
-        cart_items = order['get_total_order_quantity']
+        for i in cart:
+            product = Product.objects.get(id=i)
+            product_total_price = product.price * cart[i]['quantity']
+            order['get_total_order_price'] += product_total_price
+            order['get_total_order_quantity'] += cart[i]['quantity']
+            item = {
+                'product': {
+                    'id': product.id,
+                    'name': product.name,
+                    'price': product.price,
+                    'image': product.image,
+                },
+                'quantity': cart[i]['quantity'],
+                'get_total_items_price': product_total_price,
+            }
+            order_items.append(item)
+        
+        cart_total_quantity = order['get_total_order_quantity']
         
     context = {
         'order_items': order_items,
         'order': order,
-        'cart_items': cart_items,
+        'cart_total_quantity': cart_total_quantity,
     }
     
     if request.user.is_authenticated:
@@ -72,27 +119,47 @@ def update_order(request):
     product_id = data['productID']
     action = data['action']
     
-    customer = request.user.customer
-    product = Product.objects.get(id=product_id)
-    order, created = Order.objects.get_or_create(customer=customer, complete=False)
-    order_item, created = OrderItem.objects.get_or_create(order=order, product=product)
-    
-    if action == 'add':
-        order_item.quantity = (order_item.quantity + 1)
-    elif action == 'remove':
-        order_item.quantity = (order_item.quantity - 1)
+    if request.user.is_authenticated:
+        customer = request.user.customer
+        product = Product.objects.get(id=product_id)
+        order, created = Order.objects.get_or_create(customer=customer, complete=False)
+        order_item, created = OrderItem.objects.get_or_create(order=order, product=product)
         
-    order_item.save()
+        if action == 'add':
+            order_item.quantity = (order_item.quantity + 1)
+        elif action == 'remove':
+            order_item.quantity = (order_item.quantity - 1)
+
+        order_item.save()
+        if order_item.quantity <= 0:
+            order_item.delete()
+            
+        productQuantity = order_item.quantity
+        productPrice = floatformat(order_item.get_total_items_price, '-2g')
+        cartTotalPrice = floatformat(order.get_total_order_price, '-2g')
+        cartTotalQuantity = order.get_total_order_quantity
+    else:
+        cart = json.loads(request.COOKIES['cart'])
+        order = {'get_total_order_price':0, 'get_total_order_quantity':0}
+        for i in cart:
+            product = Product.objects.get(id=i)
+            order['get_total_order_price'] += product.price * cart[i]['quantity']
+            order['get_total_order_quantity'] += cart[i]['quantity']
     
-    if order_item.quantity <= 0:
-        order_item.delete()
+        try:
+            productQuantity = cart[product_id]['quantity']
+            productPrice = floatformat(product.price * productQuantity, '-2g')
+        except:
+            productPrice = 0
+            productQuantity = 0
+        cartTotalPrice = floatformat(order['get_total_order_price'], '-2g')
+        cartTotalQuantity = order['get_total_order_quantity']
         
     return JsonResponse({
-        'cartItems': order.get_total_order_quantity,
-        'productQuantity': order_item.quantity,
-        'productTotalPrice': floatformat(order_item.get_total_items_price, '-2'),
-        'totalOrderPrice': floatformat(order.get_total_order_price, '-2'),
-        'totalOrderQuantity': order.get_total_order_quantity,
+        'productQuantity': productQuantity,
+        'productPrice': productPrice,
+        'cartTotalPrice': cartTotalPrice,
+        'cartTotalQuantity': cartTotalQuantity,
     }, safe=False)
     
     
@@ -104,18 +171,19 @@ def place_order(request):
         customer = request.user.customer
         order, created = Order.objects.get_or_create(customer=customer, complete=False)
         total_order_price = float(data['shippingInfo']['totalOrderPrice'].replace(',', '.'))
-        order.transaction_id = transaction_id
 
-        if total_order_price == float(order.get_total_order_price):
+        if total_order_price == float(order.get_total_order_price) and total_order_price != 0:
             order.complete = True
+            order.transaction_id = transaction_id
+            ShippingAddress.objects.create(
+                customer=customer,
+                order=order,
+                address=data['shippingInfo']['address'],
+                city=data['shippingInfo']['city'],
+                country=data['shippingInfo']['country'],
+                postcode=data['shippingInfo']['postcode'],
+            )
+            
         order.save()
         
-        ShippingAddress.objects.create(
-            customer=customer,
-            order=order,
-            address=data['shippingInfo']['address'],
-            city=data['shippingInfo']['city'],
-            country=data['shippingInfo']['country'],
-            postcode=data['shippingInfo']['postcode'],
-        )
     return JsonResponse('Payment complete!', safe=False)
